Refactor MessageResponders
