+++ src/main/kotlin/me/jakejmattson/discordkt/conversations/ButtonPromptBuilder.kt
@@ -1,25 +1,20 @@
 package me.jakejmattson.discordkt.conversations
 
 import dev.kord.common.entity.ButtonStyle
-import dev.kord.common.entity.Snowflake
-import dev.kord.core.behavior.channel.createMessage
-import dev.kord.core.behavior.interaction.respondPublic
-import dev.kord.core.behavior.interaction.response.PublicMessageInteractionResponseBehavior
-import dev.kord.core.behavior.interaction.response.createPublicFollowup
-import dev.kord.core.cache.data.toData
 import dev.kord.core.entity.Message
 import dev.kord.core.entity.ReactionEmoji
 import dev.kord.core.entity.channel.MessageChannel
-import dev.kord.core.entity.interaction.followup.PublicFollowupMessage
 import dev.kord.rest.builder.message.EmbedBuilder
 import dev.kord.rest.builder.message.create.MessageCreateBuilder
 import dev.kord.rest.builder.message.create.actionRow
 import dev.kord.rest.builder.message.create.embed
 import dev.kord.x.emoji.DiscordEmoji
 import dev.kord.x.emoji.toReaction
-import kotlinx.coroutines.runBlocking
 import me.jakejmattson.discordkt.TypeContainer
 import me.jakejmattson.discordkt.commands.SlashCommandEvent
+import me.jakejmattson.discordkt.conversations.responders.ChannelResponder
+import me.jakejmattson.discordkt.conversations.responders.MessageResponder
+import me.jakejmattson.discordkt.conversations.responders.SlashResponder
 import me.jakejmattson.discordkt.dsl.uuid
 import me.jakejmattson.discordkt.extensions.toPartialEmoji
 
@@ -80,67 +75,6 @@ public open class ButtonPromptBuilder<T> {
                 }
             }
         }
-    }
-}
-
-public interface MessageResponder {
-    public val ofMessage: Message
-
-    public suspend fun respond(builder: suspend MessageCreateBuilder.() -> Unit): MessageResponder
-}
-
-public class ChannelResponder(private val channel: MessageChannel, private val message: Message) : MessageResponder {
-    public override val ofMessage: Message = message
-
-    override suspend fun respond(builder: suspend MessageCreateBuilder.() -> Unit): MessageResponder {
-        val newMessage = channel.createMessage {
-            runBlocking { builder.invoke(this@createMessage) }
-        }
-
-        return ChannelResponder(channel, newMessage)
-    }
-}
-
-public class FollowupResponder(
-    public val botResponse: PublicMessageInteractionResponseBehavior,
-    public val followupMessage: PublicFollowupMessage? = null,
-) : MessageResponder {
-    override val ofMessage: Message
-        get() = runBlocking { // FIXME don't use runBlocking
-            followupMessage?.message ?: getMessageOfBotResponse(botResponse.applicationId, botResponse.token)
-        }
-
-    private suspend fun getMessageOfBotResponse(applicationId: Snowflake, token: String): Message {
-        val messageData = botResponse.kord.rest.interaction.getInteractionResponse(applicationId, token).toData()
-
-        return Message(messageData, botResponse.kord)
-    }
-
-    override suspend fun respond(builder: suspend MessageCreateBuilder.() -> Unit): MessageResponder {
-        val newFollowupMessage = botResponse.createPublicFollowup {
-            runBlocking { builder.invoke(this@createPublicFollowup) } // FIXME don't use runBlocking
-        }
-
-        return FollowupResponder(botResponse, newFollowupMessage)
-    }
-}
-
-public class SlashResponder<T : TypeContainer>(private val event: SlashCommandEvent<T>) : MessageResponder {
-    override val ofMessage: Message
-        get() = runBlocking { getMessageOfBotResponse(event.interaction!!.applicationId, event.interaction!!.token) } // FIXME don't use runBlocking
-
-    private suspend fun getMessageOfBotResponse(applicationId: Snowflake, token: String): Message {
-        val messageData = event.discord.kord.rest.interaction.getInteractionResponse(applicationId, token).toData()
-
-        return Message(messageData, event.discord.kord)
-    }
-
-    override suspend fun respond(builder: suspend MessageCreateBuilder.() -> Unit): MessageResponder {
-        val responseBehavior = event.interaction!!.respondPublic {
-            runBlocking { builder.invoke(this@respondPublic) } // FIXME don't use runBlocking
-        }
-
-        return FollowupResponder(responseBehavior)
     }
 }
