+++ src/main/kotlin/me/jakejmattson/discordkt/conversations/responders/MessageResponder.kt
@@ -1,0 +1,10 @@
+package me.jakejmattson.discordkt.conversations.responders
+
+import dev.kord.core.entity.Message
+import dev.kord.rest.builder.message.create.MessageCreateBuilder
+
+public interface MessageResponder {
+    public val ofMessage: Message
+
+    public suspend fun respond(builder: suspend MessageCreateBuilder.() -> Unit): MessageResponder
+}
